Test package with a single source files
