--- recipes/quickfix/all/test_package/source/examples/ordermatch/test/ordermatch_ut.cpp
@@ -1,34 +1,0 @@
-/****************************************************************************
-** Copyright (c) 2001-2014
-**
-** This file is part of the QuickFIX FIX Engine
-**
-** This file may be distributed under the terms of the quickfixengine.org
-** license as defined by quickfixengine.org and appearing in the file
-** LICENSE included in the packaging of this file.
-**
-** This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
-** WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
-**
-** See http://www.quickfixengine.org/LICENSE for licensing information.
-**
-** Contact ask@quickfixengine.org if any conditions of this licensing are
-** not clear to you.
-**
-****************************************************************************/
-
-#ifdef _MSC_VER
-#pragma warning( disable : 4503 4355 4786 )
-#else
-#include "config.h"
-#endif
-
-#include "../../../src/getopt-repl.h"
-#include <UnitTest++.h>
-
-#include "OrderMatcherTestCase.cpp"
-
-int main( int argc, char** argv )
-{
-  return UnitTest::RunAllTests();
-}
